+++ everscale/contracts/utils/cell-encoder/ProxyMultiVaultCellEncoder.sol
@@ -85,7 +85,7 @@ contract ProxyMultiVaultCellEncoder {
     function encodeMultiVaultNativeSolanaEverscale(
         int8 token_wid,
         uint256 token_addr,
-        uint128 amount,
+        uint64 amount,
         int8 recipient_wid,
         uint256 recipient_addr
     ) external pure returns (TvmCell) {
@@ -125,12 +125,12 @@ contract ProxyMultiVaultCellEncoder {
         string name,
         string symbol,
         uint8 decimals,
-        uint128 amount,
+        uint64 amount,
         uint256 recipient
     ) {
         (token_wid, token_addr, name, symbol, decimals, amount, recipient) = abi.decode(
             data,
-            (int8, uint256, string, string, uint8, uint128, uint256)
+            (int8, uint256, string, string, uint8, uint64, uint256)
         );
     }
 }
